refactor: unify duplicated LLM client logic into LLMClientBase
editorial_generator and enrich_content each re-implemented the same 10 LLM
methods (~500 lines apiece): provider routing (_call_groq), the OpenAI-compatible
wrappers, Google AI / HuggingFace callers, and JSON repair/parse. Hoist them to
a single LLMClientBase in llm_client.py that both classes inherit.

Reconciliations (Rule 7 - adopt the better variant, don't average):
- _parse_json_response: keep enrich's markdown-fence stripping (editorial gains it)
- _call_google_ai_structured: keep enrich's camelCase responseMimeType/responseSchema
  (Gemini's REST API form; editorial's snake_case was likely ignored)
- log failures at warning, not error
- Preserve per-class routing via DEFAULT_MAX_TOKENS / DEFAULT_TASK_COMPLEXITY
  class attrs (editorial=800/complex, enrich=500/simple) - all call sites pass
  max_tokens explicitly but rely on the task_complexity default.

Net ~470 lines removed. Add behavior tests for the shared parse/repair logic
and the routing defaults.
