+++ scripts/editorial_generator.py
@@ -13,7 +13,6 @@ import json
 import logging
 import os
 import re
-import time
 import requests
 from dataclasses import dataclass, asdict
 from datetime import datetime
@@ -21,24 +20,11 @@ from pathlib import Path
 from typing import Dict, List, Optional
 
 from config import LLM_MIN_CALL_INTERVAL, LLM_MAX_RETRY_WAIT
-from json_utils import escape_control_chars_in_strings
-from llm_client import (
-    call_openai_compatible,
-    GROQ_SPEC,
-    OPENROUTER_SPEC,
-    OPENCODE_SPEC,
-    MISTRAL_SPEC,
-)
+from llm_client import LLMClientBase
 from editorial_renderer import generate_article_html, generate_amp_html
 from articles_index_renderer import generate_articles_index_html
 
 try:
-    from rate_limiter import (
-        get_rate_limiter,
-        check_before_call,
-        mark_provider_exhausted,
-        is_provider_exhausted,
-    )
     from shared_components import (
         build_header,
         build_footer,
@@ -47,12 +33,6 @@ try:
         get_theme_script,
     )
 except ImportError:
-    from scripts.rate_limiter import (
-        get_rate_limiter,
-        check_before_call,
-        mark_provider_exhausted,
-        is_provider_exhausted,
-    )
     from scripts.shared_components import (
         build_header,
         build_footer,
@@ -150,7 +130,7 @@ class WhyThisMatters:
     impact_areas: List[str]  # e.g., ["technology", "privacy", "business"]
 
 
-class EditorialGenerator:
+class EditorialGenerator(LLMClientBase):
     """
     Generates editorial articles and 'Why This Matters' context.
 
@@ -161,6 +141,8 @@ class EditorialGenerator:
     # enrich_content stay in sync).
     MIN_CALL_INTERVAL = LLM_MIN_CALL_INTERVAL
     MAX_RETRY_WAIT = LLM_MAX_RETRY_WAIT
+    DEFAULT_MAX_TOKENS = 800
+    DEFAULT_TASK_COMPLEXITY = "complex"
 
     def __init__(
         self,
@@ -740,507 +722,6 @@ DATE: {datetime.now().strftime('%B %d, %Y')}"""
     ) -> str:
         return generate_amp_html(article, tokens)
 
-    def _call_groq(
-        self,
-        prompt: str,
-        max_tokens: int = 800,
-        max_retries: int = 1,
-        task_complexity: str = "complex",
-    ) -> Optional[str]:
-        """
-        Call LLM API with smart provider routing based on task complexity.
-
-        For simple tasks: OpenCode (free) > Mistral (free) > Hugging Face (free) > Groq > OpenRouter > Google AI
-        For complex tasks: Mistral > Google AI > OpenRouter > OpenCode > Hugging Face > Groq
-
-        Note: Editorial defaults to 'complex' as it requires high-quality writing.
-        """
-        if task_complexity == "simple":
-            # For simple tasks, prioritize free models to save quota
-            result = self._call_opencode(prompt, max_tokens, max_retries)
-            if result:
-                return result
-
-            result = self._call_mistral(prompt, max_tokens, max_retries)
-            if result:
-                return result
-
-            result = self._call_huggingface(prompt, max_tokens, max_retries)
-            if result:
-                return result
-
-            result = self._call_groq_direct(prompt, max_tokens, max_retries)
-            if result:
-                return result
-
-            result = self._call_openrouter(prompt, max_tokens, max_retries)
-            if result:
-                return result
-
-            return self._call_google_ai(prompt, max_tokens, max_retries)
-        else:
-            # For complex tasks, prioritize higher quality models (Mistral is high quality)
-            result = self._call_mistral(prompt, max_tokens, max_retries)
-            if result:
-                return result
-
-            result = self._call_google_ai(prompt, max_tokens, max_retries)
-            if result:
-                return result
-
-            result = self._call_openrouter(prompt, max_tokens, max_retries)
-            if result:
-                return result
-
-            result = self._call_opencode(prompt, max_tokens, max_retries)
-            if result:
-                return result
-
-            result = self._call_huggingface(prompt, max_tokens, max_retries)
-            if result:
-                return result
-
-            return self._call_groq_direct(prompt, max_tokens, max_retries)
-
-    def _call_google_ai(
-        self, prompt: str, max_tokens: int = 800, max_retries: int = 1
-    ) -> Optional[str]:
-        """Call Google AI (Gemini) API - primary provider with generous free tier."""
-        if not self.google_key:
-            logger.info("No Google AI API key available, skipping to next provider")
-            return None
-
-        # Check rate limits before calling
-        rate_limiter = get_rate_limiter()
-        status = check_before_call("google")
-
-        if not status.is_available:
-            logger.warning(f"Google AI not available: {status.error}")
-            return None
-
-        if status.wait_seconds > 0:
-            logger.info(
-                f"Waiting {status.wait_seconds:.1f}s for Google AI rate limit..."
-            )
-            time.sleep(status.wait_seconds)
-
-        # Use Gemini 2.5 Flash Lite - highest RPM (10) among free models
-        model = "gemini-2.5-flash-lite"
-        url = f"https://generativelanguage.googleapis.com/v1beta/models/{model}:generateContent"
-
-        for attempt in range(max_retries):
-            try:
-                logger.info(
-                    f"Trying Google AI {model} (attempt {attempt + 1}/{max_retries})"
-                )
-                response = self.session.post(
-                    url,
-                    headers={
-                        "x-goog-api-key": self.google_key,
-                        "Content-Type": "application/json",
-                    },
-                    json={
-                        "contents": [{"parts": [{"text": prompt}]}],
-                        "generationConfig": {
-                            "maxOutputTokens": max_tokens,
-                            "temperature": 0.7,
-                        },
-                    },
-                    timeout=60,
-                )
-                response.raise_for_status()
-
-                # Update rate limiter tracking
-                rate_limiter._last_call_time["google"] = time.time()
-
-                # Parse response
-                data = response.json()
-                candidates = data.get("candidates", [])
-                if candidates:
-                    content = candidates[0].get("content", {})
-                    parts = content.get("parts", [])
-                    if parts:
-                        text = parts[0].get("text", "")
-                        if text:
-                            logger.info(f"Google AI success with {model}")
-                            return text
-
-            except requests.exceptions.HTTPError as e:
-                if response.status_code == 429:
-                    # Check if this is a quota exhaustion (daily limit) vs temporary rate limit
-                    try:
-                        error_data = response.json()
-                        error_msg = str(error_data).lower()
-                        if (
-                            "quota" in error_msg
-                            or "exhausted" in error_msg
-                            or "daily" in error_msg
-                        ):
-                            # This is a quota exhaustion - mark provider as exhausted
-                            mark_provider_exhausted("google", "daily quota exceeded")
-                            return None
-                    except (ValueError, KeyError) as parse_err:
-                        logger.debug(f"Could not parse 429 body for quota check: {parse_err}")
-
-                    # Temporary rate limit - wait and retry
-                    retry_after = response.headers.get("Retry-After", "10")
-                    try:
-                        wait_time = min(float(retry_after), self.MAX_RETRY_WAIT)
-                    except ValueError:
-                        wait_time = self.MAX_RETRY_WAIT
-                    logger.warning(
-                        f"Google AI rate limited, waiting {wait_time}s (attempt {attempt + 1}/{max_retries})"
-                    )
-                    time.sleep(wait_time)
-                    continue
-                logger.error(f"Google AI failed: {e}")
-                return None
-            except (requests.RequestException, json.JSONDecodeError, KeyError, ValueError, AttributeError) as e:
-                logger.error(f"Google AI failed: {e}")
-                return None
-
-        logger.warning("Google AI: Max retries exceeded")
-        return None
-
-    def _call_google_ai_structured(
-        self, prompt: str, schema: dict, max_tokens: int = 2000, max_retries: int = 1
-    ) -> Optional[Dict]:
-        """
-        Call Google AI with structured output (guaranteed valid JSON).
-
-        Uses Gemini's response_mime_type and response_schema to ensure
-        the response matches the provided JSON schema.
-        """
-        if not self.google_key:
-            logger.info("No Google AI API key available, skipping structured output")
-            return None
-
-        # Check rate limits before calling
-        rate_limiter = get_rate_limiter()
-        status = check_before_call("google")
-
-        if not status.is_available:
-            logger.warning(f"Google AI not available: {status.error}")
-            return None
-
-        if status.wait_seconds > 0:
-            logger.info(
-                f"Waiting {status.wait_seconds:.1f}s for Google AI rate limit..."
-            )
-            time.sleep(status.wait_seconds)
-
-        # Use Gemini 2.5 Flash Lite with structured output
-        model = "gemini-2.5-flash-lite"
-        url = f"https://generativelanguage.googleapis.com/v1beta/models/{model}:generateContent"
-
-        for attempt in range(max_retries):
-            try:
-                logger.info(
-                    f"Trying Google AI {model} with structured output (attempt {attempt + 1}/{max_retries})"
-                )
-                response = self.session.post(
-                    url,
-                    headers={
-                        "x-goog-api-key": self.google_key,
-                        "Content-Type": "application/json",
-                    },
-                    json={
-                        "contents": [{"parts": [{"text": prompt}]}],
-                        "generationConfig": {
-                            "maxOutputTokens": max_tokens,
-                            "temperature": 0.7,
-                            "response_mime_type": "application/json",
-                            "response_schema": schema,
-                        },
-                    },
-                    timeout=90,  # Longer timeout for structured output
-                )
-                response.raise_for_status()
-
-                # Update rate limiter tracking
-                rate_limiter._last_call_time["google"] = time.time()
-
-                # Parse response - should be valid JSON
-                data = response.json()
-                candidates = data.get("candidates", [])
-                if candidates:
-                    content = candidates[0].get("content", {})
-                    parts = content.get("parts", [])
-                    if parts:
-                        text = parts[0].get("text", "")
-                        if text:
-                            try:
-                                result = json.loads(text)
-                                logger.info(
-                                    f"Google AI structured output success with {model}"
-                                )
-                                return result
-                            except json.JSONDecodeError as e:
-                                # Shouldn't happen with structured output, but fallback to repair
-                                logger.warning(
-                                    f"Structured output JSON parse error (unexpected): {e}"
-                                )
-                                repaired = self._repair_json(text)
-                                return json.loads(repaired)
-
-            except requests.exceptions.HTTPError as e:
-                if response.status_code == 429:
-                    # Check if this is a quota exhaustion (daily limit) vs temporary rate limit
-                    try:
-                        error_data = response.json()
-                        error_msg = str(error_data).lower()
-                        if (
-                            "quota" in error_msg
-                            or "exhausted" in error_msg
-                            or "daily" in error_msg
-                        ):
-                            # This is a quota exhaustion - mark provider as exhausted
-                            mark_provider_exhausted("google", "daily quota exceeded")
-                            return None
-                    except (ValueError, KeyError) as parse_err:
-                        logger.debug(f"Could not parse 429 body for quota check: {parse_err}")
-
-                    # Temporary rate limit - wait and retry
-                    retry_after = response.headers.get("Retry-After", "10")
-                    try:
-                        wait_time = min(float(retry_after), self.MAX_RETRY_WAIT)
-                    except ValueError:
-                        wait_time = self.MAX_RETRY_WAIT
-                    logger.warning(
-                        f"Google AI rate limited, waiting {wait_time}s (attempt {attempt + 1}/{max_retries})"
-                    )
-                    time.sleep(wait_time)
-                    continue
-                logger.error(f"Google AI structured output failed: {e}")
-                return None
-            except (requests.RequestException, json.JSONDecodeError, KeyError, ValueError, AttributeError) as e:
-                logger.error(f"Google AI structured output failed: {e}")
-                return None
-
-        logger.warning("Google AI structured output: Max retries exceeded")
-        return None
-
-    def _call_openrouter(
-        self, prompt: str, max_tokens: int = 800, max_retries: int = 1
-    ) -> Optional[str]:
-        """Call OpenRouter API with free models (primary)."""
-        result = call_openai_compatible(
-            OPENROUTER_SPEC, self.openrouter_key, prompt, max_tokens, max_retries,
-            self.session, self.MAX_RETRY_WAIT,
-        )
-        return result
-
-    def _call_groq_direct(
-        self, prompt: str, max_tokens: int = 800, max_retries: int = 1
-    ) -> Optional[str]:
-        """Call Groq API directly (fallback)."""
-        timing = [self._last_call_time]
-        result = call_openai_compatible(
-            GROQ_SPEC, self.groq_key, prompt, max_tokens, max_retries,
-            self.session, self.MAX_RETRY_WAIT, self.MIN_CALL_INTERVAL, timing,
-        )
-        self._last_call_time = timing[0]
-        return result
-
-    def _call_opencode(
-        self, prompt: str, max_tokens: int = 800, max_retries: int = 1
-    ) -> Optional[str]:
-        """Call OpenCode API with free models (glm-4.7-free, minimax-m2.1-free)."""
-        opencode_key = os.getenv("OPENCODE_API_KEY")
-        timing = [self._last_call_time]
-        result = call_openai_compatible(
-            OPENCODE_SPEC, opencode_key, prompt, max_tokens, max_retries,
-            self.session, self.MAX_RETRY_WAIT, self.MIN_CALL_INTERVAL, timing,
-        )
-        self._last_call_time = timing[0]
-        return result
-
-    def _call_huggingface(
-        self, prompt: str, max_tokens: int = 800, max_retries: int = 1
-    ) -> Optional[str]:
-        """Call Hugging Face Inference API with free models."""
-        huggingface_key = os.getenv("HUGGINGFACE_API_KEY")
-        if not huggingface_key:
-            return None
-
-        # Check rate limits before calling
-        rate_limiter = get_rate_limiter()
-        status = check_before_call("huggingface")
-
-        if not status.is_available:
-            logger.warning(f"Hugging Face not available: {status.error}")
-            return None
-
-        if status.wait_seconds > 0:
-            logger.info(
-                f"Waiting {status.wait_seconds:.1f}s for Hugging Face rate limit..."
-            )
-            time.sleep(status.wait_seconds)
-
-        # Proactive rate limiting
-        elapsed = time.time() - self._last_call_time
-        if elapsed < self.MIN_CALL_INTERVAL:
-            time.sleep(self.MIN_CALL_INTERVAL - elapsed)
-
-        # Free models to try in order (7B models work well on free tier)
-        free_models = [
-            "mistralai/Mistral-7B-Instruct-v0.3",
-            "Qwen/Qwen2.5-7B-Instruct",
-            "microsoft/Phi-3-mini-4k-instruct",
-        ]
-
-        for model in free_models:
-            for attempt in range(max_retries):
-                try:
-                    self._last_call_time = time.time()
-                    logger.info(
-                        f"Trying Hugging Face {model} (attempt {attempt + 1}/{max_retries})"
-                    )
-                    response = self.session.post(
-                        f"https://api-inference.huggingface.co/models/{model}",
-                        headers={
-                            "Authorization": f"Bearer {huggingface_key}",
-                            "Content-Type": "application/json",
-                        },
-                        json={
-                            "inputs": prompt,
-                            "parameters": {
-                                "max_new_tokens": max_tokens,
-                                "temperature": 0.7,
-                                "return_full_text": False,
-                            },
-                        },
-                        timeout=60,
-                    )
-                    response.raise_for_status()
-
-                    # Update rate limiter from response headers
-                    rate_limiter.update_from_response_headers(
-                        "huggingface", dict(response.headers)
-                    )
-                    rate_limiter._last_call_time["huggingface"] = time.time()
-
-                    result = response.json()
-                    if isinstance(result, list) and len(result) > 0:
-                        text = result[0].get("generated_text", "")
-                        if text:
-                            logger.info(f"Hugging Face success with {model}")
-                            return text
-
-                except requests.exceptions.HTTPError as e:
-                    if response.status_code == 429:
-                        retry_after = response.headers.get("Retry-After", "10")
-                        try:
-                            wait_time = min(float(retry_after), self.MAX_RETRY_WAIT)
-                        except ValueError:
-                            wait_time = self.MAX_RETRY_WAIT
-                        logger.warning(
-                            f"Hugging Face rate limited, waiting {wait_time}s (attempt {attempt + 1}/{max_retries})"
-                        )
-                        time.sleep(wait_time)
-                        continue
-                    elif response.status_code == 503:
-                        # Model is loading, wait and retry
-                        logger.warning(
-                            f"Hugging Face model {model} is loading, waiting {self.MAX_RETRY_WAIT}s..."
-                        )
-                        time.sleep(self.MAX_RETRY_WAIT)
-                        continue
-                    logger.warning(f"Hugging Face API error with {model}: {e}")
-                    break  # Try next model
-                except (requests.RequestException, json.JSONDecodeError, KeyError, ValueError, AttributeError) as e:
-                    logger.warning(f"Hugging Face API error with {model}: {e}")
-                    break  # Try next model
-
-        logger.warning("All Hugging Face models failed")
-        return None
-
-    def _call_mistral(
-        self, prompt: str, max_tokens: int = 800, max_retries: int = 1
-    ) -> Optional[str]:
-        """Call Mistral AI API - high quality free tier models."""
-        mistral_key = os.getenv("MISTRAL_API_KEY")
-        timing = [self._last_call_time]
-        result = call_openai_compatible(
-            MISTRAL_SPEC, mistral_key, prompt, max_tokens, max_retries,
-            self.session, self.MAX_RETRY_WAIT, self.MIN_CALL_INTERVAL, timing,
-        )
-        self._last_call_time = timing[0]
-        return result
-
-    def _repair_json(self, json_str: str) -> str:
-        """Attempt to repair common JSON formatting issues from LLM output."""
-        # Fix missing commas between elements (common LLM error)
-        # Pattern: }" followed by whitespace and then "{ or "[
-        json_str = re.sub(r'"\s*\n\s*"', '",\n"', json_str)
-        json_str = re.sub(r"}\s*\n\s*{", "},\n{", json_str)
-        json_str = re.sub(r"]\s*\n\s*\[", "],\n[", json_str)
-        json_str = re.sub(r'"\s*\n\s*{', '",\n{', json_str)
-        json_str = re.sub(r'}\s*\n\s*"', '},\n"', json_str)
-        json_str = re.sub(r'"\s*\n\s*\[', '",\n[', json_str)
-        json_str = re.sub(r']\s*\n\s*"', '],\n"', json_str)
-
-        # Fix missing comma after value before next key
-        # Pattern: "value" (whitespace) "key":
-        json_str = re.sub(r'"\s+("[\w]+"\s*:)', r'", \1', json_str)
-
-        # Fix trailing commas before closing brackets
-        json_str = re.sub(r",\s*}", "}", json_str)
-        json_str = re.sub(r",\s*]", "]", json_str)
-
-        return json_str
-
-    def _parse_json_response(self, response: Optional[str]) -> Optional[Dict]:
-        """Parse JSON from LLM response."""
-        if not response:
-            return None
-
-        try:
-            # Try to find JSON in response
-            json_match = re.search(r"\{.*\}", response, re.DOTALL)
-            if json_match:
-                json_str = json_match.group()
-                # First, try parsing as-is
-                try:
-                    return json.loads(json_str)
-                except json.JSONDecodeError:
-                    pass
-
-                # Try repairing common JSON issues (missing commas, etc.)
-                try:
-                    repaired = self._repair_json(json_str)
-                    return json.loads(repaired)
-                except json.JSONDecodeError:
-                    pass
-
-                # Escape raw control characters that appear inside quoted
-                # strings (a common defect in LLM JSON output).
-                try:
-                    sanitized = escape_control_chars_in_strings(json_str)
-                    return json.loads(sanitized)
-                except json.JSONDecodeError:
-                    pass
-
-                # Try repair + escape combination
-                try:
-                    repaired = self._repair_json(json_str)
-                    sanitized = escape_control_chars_in_strings(repaired)
-                    return json.loads(sanitized)
-                except json.JSONDecodeError:
-                    pass
-
-                # Last resort: strip all control chars except structural whitespace
-                try:
-                    stripped = re.sub(r"[\x00-\x09\x0b\x0c\x0e-\x1f]", " ", json_str)
-                    repaired = self._repair_json(stripped)
-                    return json.loads(repaired)
-                except json.JSONDecodeError:
-                    pass
-
-        except (json.JSONDecodeError, ValueError) as e:
-            logger.warning(f"JSON parse error: {e}")
-
-        return None
 
     def _sanitize_slug(self, slug: str) -> str:
         """Sanitize slug for URL usage."""
@@ -1421,7 +902,6 @@ DATE: {datetime.now().strftime('%B %d, %Y')}"""
 
         logger.info(f"Generated enhanced articles index with {len(articles)} articles")
         return html
-
 
 
 if __name__ == "__main__":
